+++ Domain/Entities/EntityCadastrosVeiculo.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Chassi { get; set; }
         public string Tipo { get; set; }
-        public int NumeroPassageiros { get; set; }
+        public Byte NumeroPassageiros { get; set; }
         public string Cor { get; set; }     
         
     }
